Move ecdsa python implementation into separate backend. Key.to_pem returns six.binary_type - str on PY2, bytes on PY3.
