+++ jose/backends/__init__.py
@@ -1,5 +1,10 @@
 
 try:
-    from .pycrypto_backend import RSAKey
+    from jose.backends.pycrypto_backend import RSAKey
 except ImportError:
-    from .cryptography_backend import CryptographyRSAKey as RSAKey
+    from jose.backends.cryptography_backend import CryptographyRSAKey as RSAKey
+
+try:
+    from jose.backends.ecdsa_backend import ECKey
+except ImportError:
+    from jose.backends.cryptography_backend import CryptographyECKey as ECKey
